Allow the DebugOverlay to appear on a specific screen and display only touches of a specific digitizer
